Docs: Refactor NotePane and Update DG
