+++ src/main/java/seedu/notor/ui/MainWindow.java
@@ -29,7 +29,6 @@ import seedu.notor.ui.note.GeneralNoteWindow;
 import seedu.notor.ui.note.GroupNoteWindow;
 import seedu.notor.ui.note.NoteWindow;
 import seedu.notor.ui.note.PersonNoteWindow;
-import seedu.notor.ui.view.ViewPanel;
 
 /**
  * The Main Window. Provides the basic application layout containing
@@ -55,7 +54,7 @@ public class MainWindow extends UiPart<Stage> {
     private StackPane commandBoxPlaceholder;
 
     @FXML
-    private StackPane notePanePlaceholder;
+    private StackPane generalNotePlaceholder;
 
     @FXML
     private MenuItem helpMenuItem;
@@ -145,8 +144,8 @@ public class MainWindow extends UiPart<Stage> {
         CommandBox commandBox = new CommandBox(this::executeCommand);
         commandBoxPlaceholder.getChildren().add(commandBox.getRoot());
 
-        ViewPanel viewPane = new ViewPanel(logic.getNotor().getNote());
-        notePanePlaceholder.getChildren().add(viewPane.getRoot());
+        GeneralNote viewPane = new GeneralNote(logic.getNotor().getNote());
+        generalNotePlaceholder.getChildren().add(viewPane.getRoot());
     }
 
     /**
@@ -198,7 +197,7 @@ public class MainWindow extends UiPart<Stage> {
     }
 
     private void handleNote(Notor notor, Logic logic) {
-        NoteWindow noteWindow = new GeneralNoteWindow(notor, logic, resultDisplay, notePanePlaceholder);
+        NoteWindow noteWindow = new GeneralNoteWindow(notor, logic, resultDisplay, generalNotePlaceholder);
         manageNoteWindow(noteWindow);
     }
