+++ src/main/resources/view/MainWindow.fxml
@@ -66,7 +66,7 @@
               <items>
                   <VBox alignment="BOTTOM_CENTER" styleClass="splitpane-background">
                      <children>
-                        <StackPane fx:id="notePanePlaceholder" maxHeight="1.7976931348623157E308" />
+                        <StackPane fx:id="generalNotePlaceholder" maxHeight="1.7976931348623157E308" />
                         <StackPane fx:id="commandBoxPlaceholder" styleClass="pane-with-border" />
                         <StackPane fx:id="resultDisplayPlaceholder" minHeight="-Infinity" prefHeight="100.0" VBox.vgrow="ALWAYS" />
                      </children>
